split redirections functions
