+++ include/minishell_exec.h
@@ -6,7 +6,7 @@
 /*   By: lusavign <lusavign@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/18 18:08:01 by racoutte          #+#    #+#             */
-/*   Updated: 2025/02/21 17:41:51 by lusavign         ###   ########.fr       */
+/*   Updated: 2025/02/21 19:32:11 by lusavign         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,13 +56,16 @@ int					ft_cd(t_exec *ex, t_env *env);
 int					ft_env(t_exec *ex, t_env *env);
 int					ft_export(t_exec *exec, t_env **env);
 int					ft_exit(t_exec *ex, t_env *env, int *std_dup);
-long				*get_exit_status(void);
-void				modify_value_exit_code(long code);
+int					check_if_cmd_has_arg(char **arg);
 int					check_if_var_name_is_valid(char *arg);
+
+long				*get_exit_status(void);
+
 char				*extract_key_name(char *arg);
 char				*extract_var_value(char *arg);
+
+void				modify_value_exit_code(long code);
 void				modify_value(char *new_value, t_env *env_var);
-int					check_if_cmd_has_arg(char **arg);
 
 // EXEC UTILS//
 int					ft_strcmp(char *s1, char *s2);
@@ -71,19 +74,29 @@ int					count_command(t_exec *ex);
 int					has_pipe(t_exec *ex);
 int					has_redir(t_exec *ex);
 int					has_heredoc(t_exec *ex);
+
 char				*ft_strndup(const char *s, size_t n);
+
 void				ft_init(t_exec *ex, int *std_dup);
 void    			restore_fds(int *std_dup);
 void				ft_close_fds(int fd);
 void				ft_close_fd(int *pipefd);
+void    			ft_open(t_exec *ex, int *fd_in);
+void				ft_open_heredocs(t_exec *ex, int pipefd);
 
 // EXEC//
-char				**put_env_in_ar(t_env *envp);
-char				*is_path_exec(char *cmd, char **full_paths);
-char				*get_path(t_env *env, char *cmd);
 int					exec_builtin(t_exec *ex, t_env **env, int *std_dup);
 int					is_builtin(t_exec *ex);
 int					handle_heredoc(t_exec *ex);
+
+char				**put_env_in_ar(t_env *envp);
+char				*is_path_exec(char *cmd, char **full_paths);
+char				*get_path(t_env *env, char *cmd);
+
+void				redir_out(t_exec *ex);
+void				redir_in(t_exec *ex, int *fd_in);
+void				handle_redir(t_exec *ex);
+void				handle_redir_in_pipe(t_exec *ex, int pipefd);
 void				ft_fork(t_exec *cmd, t_env **env, int *std_dup);
 void    			ft_process(t_env **env, t_exec *ex);
