+++ src/exec/heredoc.c
@@ -6,11 +6,39 @@
 /*   By: lusavign <lusavign@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/10 15:37:43 by lusavign          #+#    #+#             */
-/*   Updated: 2025/02/21 15:08:40 by lusavign         ###   ########.fr       */
+/*   Updated: 2025/02/21 18:21:18 by lusavign         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
+
+void    ft_open_heredocs(t_exec *ex, int pipefd)
+{
+    t_exec  *current = ex;
+    int     fd_in;
+
+    fd_in = pipefd;
+    while (current)
+    {
+        if (current->type == TOKEN_REDIR_HEREDOC)
+        {
+            if (fd_in > 2)
+                ft_close_fds(fd_in);
+            fd_in = handle_heredoc(current);
+            if (fd_in < 0)
+            {
+                perror("Error handling heredoc");
+                return;
+            }
+        }
+        current = current->next;
+    }
+    if (fd_in > 2)
+    {
+        dup2(fd_in, STDIN_FILENO);
+        ft_close_fds(fd_in);
+    }
+}
 
 char	*generate_heredoc_name(char	*heredoc)
 {
